requirements.txt
altgraph==0.17.4
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
packaging==23.2
pillow==10.2.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
requests==2.31.0
urllib3==2.2.1
